Change PickUpItem spawn logic and adding healing logic (connect between PlayerController and ResourceManager)

+++ Assets/Scripts/Managers/PickUpItemManager/AttributesSet.cs
@@ -3,7 +3,7 @@ using System;
 [Serializable]
 public struct AttributesSet
 {
-    public int Gold;
-    public int Heal;
-    public int Mana;
+    public float Gold;
+    public float Heal;
+    public float Mana;
 }
